+++ src/backend/src/main/java/com/rjb/hobby_tracker/roles/RoleDTO.java
@@ -1,0 +1,16 @@
+package com.rjb.hobby_tracker.roles;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@Getter
+@NoArgsConstructor
+@Setter
+public class RoleDTO {
+    private Integer id;
+
+    private String name;
+}