+++ src/backend/src/main/java/com/rjb/hobby_tracker/roles/RoleRepository.java
@@ -12,4 +12,12 @@ public interface RoleRepository extends JpaRepository<RoleEntity, Integer> {
      * @return Optional of role
      */
     public Optional<RoleEntity> findByName(String name);
+
+    /**
+     * Determine if role exists by provided name. 
+     * 
+     * @param name
+     * @return <i>true</i> if role with name exists, <i>false</i> otherwise
+     */
+    public boolean existsByName(String name);
 }