+++ src/backend/src/main/java/com/rjb/hobby_tracker/users/UserEntity.java
@@ -67,7 +67,5 @@ public class UserEntity implements UserDetails {
         return privileges;
       }).flatMap(List::stream)
       .collect(Collectors.toList());
-
-      // return getRoles().stream().map(r -> new SimpleGrantedAuthority(r.getName())).toList();
     }
 }